Move GUI/login components to subdirectory
